Prepare campaign for C21_WMDE_Test_17
 Use VAR of C21_WMDE_Test_16 (Multistep form)

 Change progress bar to show "n days left" (using custom desktop
 progress bar).
