+++ desktop/components/ui/ProgressBar.jsx
@@ -18,7 +18,7 @@ export const AmountToShowOnRight = Object.freeze( {
  *
  * @type {boolean}
  */
-const IS_LATE_PROGRESS = false;
+const IS_LATE_PROGRESS = true;
 
 export default class ProgressBar extends Component {
 	static contextType = TranslationContext;
@@ -81,7 +81,6 @@ export default class ProgressBar extends Component {
 	// eslint-disable-next-line no-unused-vars
 	progressAnimationEnded = ( e ) => {
 		this.setState( { animation: ENDED } );
-
 	};
 
 	startAnimation() {
@@ -95,16 +94,13 @@ export default class ProgressBar extends Component {
 	}
 
 	calculateWidth() {
-		return ( this.props.donationAmount * 100 ) / this.props.goalDonationSum;
+		return Math.min( ( this.props.donationAmount * 100 ) / this.props.goalDonationSum, 100 );
 	}
 
 	render( props, state, context ) {
 		const Translations = context;
 		const getMillion = n => this.props.formatters.millionFormatter( n / 1000000 );
 		const getDaysLeft = daysLeft => {
-			if ( props.daysLeft > 14 ) {
-				return '';
-			}
 			return Translations[ 'prefix-days-left' ] +
 				' ' + daysLeft + ' ' +
 				( daysLeft === 1 ? Translations[ 'day-singular' ] : Translations[ 'day-plural' ] ) + ' ' +
