+++ desktop/form_items.js
@@ -1,4 +1,4 @@
-import FormItemsBuilder, { Intervals, PaymentMethods } from '../shared/components/ui/form/FormItemsBuilder';
+import FormItemsBuilder, { Intervals, PaymentMethods, AddressType } from './components/ui/form/FormItemsBuilder';
 
 export function createFormItems( translations, amountFormatter ) {
 	const builder = new FormItemsBuilder( translations, amountFormatter );
@@ -14,6 +14,10 @@ export function createFormItems( translations, amountFormatter ) {
 		PaymentMethods.BANK_TRANSFER,
 		PaymentMethods.DIRECT_DEBIT,
 		PaymentMethods.CREDIT_CARD
+	);
+	builder.setAddressType(
+		AddressType.FULL,
+		AddressType.NO
 	);
 	return builder.getItems();
 }
