--- desktop/form_items_var.js
@@ -1,23 +1,0 @@
-import FormItemsBuilder, { Intervals, PaymentMethods, AddressType } from './components/ui/form/FormItemsBuilder';
-
-export function createFormItems( translations, amountFormatter ) {
-	const builder = new FormItemsBuilder( translations, amountFormatter );
-	builder.setIntervals(
-		Intervals.ONCE,
-		Intervals.MONTHLY,
-		Intervals.QUARTERLY,
-		Intervals.YEARLY
-	);
-	builder.setAmounts( 5, 10, 20, 25, 50, 100 );
-	builder.setPaymentMethods(
-		PaymentMethods.PAYPAL,
-		PaymentMethods.BANK_TRANSFER,
-		PaymentMethods.DIRECT_DEBIT,
-		PaymentMethods.CREDIT_CARD
-	);
-	builder.setAddressType(
-		AddressType.FULL,
-		AddressType.NO
-	);
-	return builder.getItems();
-}
