+++ desktop/styles/styles.pcss
@@ -4,13 +4,14 @@
 @import '../../shared/components/BannerTransition.pcss';
 
 @import './Banner.pcss';
-@import 'DonationForm.pcss';
+@import './DonationForm.pcss';
+@import './DonationFormSteps.pcss';
 @import './SelectGroup.pcss';
 @import './Footer.pcss';
 @import './Infobox.pcss';
-@import 'BannerText.pcss';
+@import './BannerText.pcss';
 @import './ProgressBar.pcss';
-@import 'Slider.pcss';
+@import './Slider.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
