+++ shared/global_banner_settings.js
@@ -1,13 +1,13 @@
 // TODO When all banners use `createCampaignParameters`, rename the properties and use this object directly.
 module.exports = {
 	'goalDonationSum': 9200000,
-	'donations-date-base': '2021-10-14',
-	'donations-collected-base': 4434000,
-	'donators-base': 35814,
-	'appr-donations-per-minute': 104,
-	'appr-donators-per-minute': 4.805,
+	'donations-date-base': '2021-11-12',
+	'donations-collected-base': 2000000,
+	'donators-base': 72200,
+	'appr-donations-per-minute': 100,
+	'appr-donators-per-minute': 4,
 	'impressions-per-day-in-million': 6,
-	'campaign-start-date': '2021-10-13',
+	'campaign-start-date': '2021-11-03',
 	'campaign-end-date': '2021-12-31',
 	'banner-close-track-ratio': 1
 };
